requirements.txt
fastapi==0.110.0
uvicorn[standard]==0.29.0
starlette==0.36.3
pydantic==2.6.4
gensim==4.3.2
numpy==1.26.4
httpx==0.27.2
scipy==1.10.1
requests==2.32.3
python-dotenv==1.2.1
sqlalchemy==2.0.28
asyncpg==0.29.0
passlib[bcrypt]==1.7.4
python-jose[cryptography]==3.3.0
python-multipart==0.0.9